Finish search API support; improve JSON deserialization error message; add OpenAPI annotations
